Moved cosmos query test to azure cosmos module
