--- data-document-processor/src/test/groovy/io/micronaut/data/document/processor/BuildCosmosQuerySpec.groovy
@@ -1,23 +1,0 @@
-package io.micronaut.data.document.processor
-
-class BuildCosmosQuerySpec extends AbstractDataSpec {
-
-    void "test cosmos repo"() {
-        given:
-        def repository = buildRepository('test.CosmosBookRepository', """
-import io.micronaut.data.cosmos.annotation.CosmosRepository;
-import io.micronaut.data.document.tck.entities.Book;
-import java.util.Optional;
-@CosmosRepository
-interface CosmosBookRepository extends GenericRepository<Book, String> {
-    Optional<Book> findById(String id);
-}
-"""
-        )
-
-        when:
-        String q = TestUtils.getQuery(repository.getRequiredMethod("findById", String))
-        then:
-        q == "SELECT book_.id,book_.authorId,book_.title,book_.totalPages,book_.publisherId,book_.lastUpdated,book_.created FROM book book_ WHERE (book_.id = @p1)"
-    }
-}
